DIDSubjectIdentifier member is now 'url'
per change to the PR prior to merge into the Subject Identifiers draft.

As such, URLs here can be relative (no scheme), and have a wide variety
of string values and formats.  Rather than try to validate that the string
is a valid URL, simply test for the common cases of missing, null, empty string.